refactor: 전체 도메인 idx값 Long 타입으로 수정
- PK 타입 Integer → Long으로 변경
- 엔티티 idx 필드 Integer → Long으로 통일
- repository.findById(idx) 호출부 타입 일치 처리
- 그외 필드들 Integer로 통일.

+++ src/main/java/kr/co/itid/cms/config/security/JwtTokenProvider.java
@@ -92,7 +92,7 @@ public class JwtTokenProvider {
     public Map<String, Object> getClaims(Member member) {
         Map<String, Object> claims = new HashMap<>();
 
-        int idx = member.getIdx();
+        Long idx = member.getIdx();
         int userLevel = member.getUserLevel();
         String userName = member.getUserName();
