+++ src/main/java/kr/co/itid/cms/entity/cms/core/member/Member.java
@@ -17,7 +17,7 @@ public class Member {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idx;
+    private Long idx;
 
     @Column(name = "user_id", nullable = false)
     private String userId;
